Refactored ShipmentController to use ObjectResults instead of exceptions, and to use early exit pattern for variables that could be null.

+++ SmartWarehouseManagement.Server/Controllers/ShipmentController.cs
@@ -25,16 +25,22 @@ namespace SmartWarehouseManagement.Server.Controllers
         [HttpGet("{id}")]
         public ActionResult<Shipment> GetShipment(int id)
         {
-            return _dbContext.Shipments.Include(s => s.Order)
-                                       .FirstOrDefault(s => s.Id == id)
-                                       ?? throw new InvalidOperationException("No shipment with id: " + id + " exists.");
+            var shipment = _dbContext.Shipments.Include(s => s.Order)
+                .FirstOrDefault(s => s.Id == id);
+
+            if (shipment == null) {
+                return BadRequest($"No shipment with id: {id} exists");
+            }
+
+            return shipment;
         }
 
         [HttpPost]
-        public ActionResult<Shipment> CreateShipment(Shipment shipment) { 
-
-            Order order = _dbContext.Orders.Find(shipment.Order.Id) ?? 
-                throw new InvalidOperationException("No order with id: " + shipment.Order.Id + " exists.");
+        public ActionResult<Shipment> CreateShipment(Shipment shipment) {
+            if (_dbContext.Orders.Find(shipment.Order.Id) is not Order order)
+            {
+                return BadRequest($"No order with an id of {shipment.Order.Id} exists");
+            }
 
             if (order.Shipped)
             {
@@ -51,12 +57,21 @@ namespace SmartWarehouseManagement.Server.Controllers
         }
 
         [HttpDelete("{id}")]
-        public IActionResult DeleteShipment(int id) { 
-            Shipment shipmentToDelete = _dbContext.Shipments.Include(s => s.Order)
-                                                            .FirstOrDefault(s => s.Id == id) 
-                                                            ?? throw new InvalidOperationException("No shipment with id: " + id + " exists.");
-            
-            _dbContext.Orders.Find(shipmentToDelete.Order.Id).Shipped = false;
+        public IActionResult DeleteShipment(int id) {
+            var shipmentToDelete = _dbContext.Shipments
+                .Include(s => s.Order)
+                .FirstOrDefault(s => s.Id == id);
+
+            if (shipmentToDelete == null) {
+                return BadRequest($"No shipment with id: {id} exists.");
+            }
+
+            if (_dbContext.Orders.Find(shipmentToDelete.Order.Id) is not Order order)
+            {
+                return BadRequest($"No order with an id of {shipmentToDelete.Order.Id} exists");
+            }
+
+            order.Shipped = false;
 
             _dbContext.Shipments.Remove(shipmentToDelete);
             _dbContext.SaveChanges();
